fix(admin): cascading errors and updated typings
