+++ src/bp/ui-admin/src/Pages/Server/Languages/Language.tsx
@@ -54,7 +54,6 @@ const Language: FC<Props> = props => {
   const loadLanguage = async () => {
     setLoading(true)
     try {
-      // @ts-ignore
       await api.getSecured({ timeout: 10000 }).post(`/admin/languages/${props.language.code}/load`)
     } catch (err) {
       console.log('error loading model')
